+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add options for SFTP export ([#223](../../pull/223))
 - Support for scanning image labels with OCR ([#233](../../pull/233))
 - Allow changing field in import and export reports ([#244](../../pull/244))
+- Allow redacting and approving lists of images ([#247](../../pull/247))
 
 ### Improvements
 - Added more logging during export ([#222](../../pull/222))